requirements.txt
numpy
Pillow
streamlit==0.79.0
opencv-python-headless==4.2.0.32
